refact: replaces int primitive data type by Long in entity classes and repository

+++ src/main/java/com/abutua/productbackend/models/Category.java
@@ -1,5 +1,7 @@
 package com.abutua.productbackend.models;
 
+import java.util.function.LongConsumer;
+
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -12,11 +14,11 @@ public class Category {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     
     private String name;
 
-    public Category(int id, String name) {
+    public Category(Long id, String name) {
         this.id = id;
         this.name = name;
     }
@@ -24,11 +26,11 @@ public class Category {
     public Category() {
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
